pkg - Push to Launchpad using SFTP (finalized)
---

We can finally push to Launchpad using SFTP, and this is beneficial for
Nitrocid KS. However, this is not project-independent.

We have tested this code, and pushing to Launchpad went fine. Because we
already have 1.0.1 in the PPA, we got a rejection e-mail, but it shouldn't
be a problem.

Rejection e-mail states:

Rejected:
File bskyid_1.0.1.orig.tar.xz already exists in BlueSky ID, but uploaded
version has different contents. See more information about this error in
https://help.launchpad.net/Packaging/UploadErrors.

---

Type: pkg
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1